Reflactor Collection.vue
