+++ app/javascript/application.js
@@ -36,11 +36,11 @@ import OrganizationDetail from "./components/organizations/OrganizationDetail.vu
 import OnPremiseFooter from "./components/OnPremiseFooter.vue"
 import NewApplicationSpace from "./components/application_spaces/NewApplicationSpace.vue"
 import ProfileRepoList from "./components/shared/ProfileRepoList.vue"
-import ProfileRepoLikesList from "./components/shared/ProfileRepoLikesList.vue"
 import RepoCards from "./components/shared/RepoCards.vue"
 import NewCode from "./components/codes/NewCode.vue"
 import CodeDetail from "./components/codes/CodeDetail.vue"
 import SvgIcon from "./components/shared/SvgIcon.vue"
+import Collection from "./components/shared/Collection.vue"
 
 const pinia = createPinia()
 
@@ -64,11 +64,11 @@ const app = createApp({
     OnPremiseFooter,
     NewApplicationSpace,
     ProfileRepoList,
-    ProfileRepoLikesList,
     OrganizationSettings,
     RepoCards,
     NewCode,
-    CodeDetail
+    CodeDetail,
+    Collection
   },
   provide:{
     defaultTags: DEFAULT_TAGS,
