+++ app/views/profile/likes.html.erb
@@ -1,5 +1,5 @@
-<div class="flex justify-center lg:flex-col my-[24px] rounded-[8px] md:px-[50px] sm:px-[20px] max-w-[1280px] m-auto bg-white">
-  <div class="md:mb-[24px] bg-[#FAFAFA] sm:bg-white md:bg-white">
+<div class="flex justify-center lg:flex-col rounded-[8px] md:px-[50px] sm:px-[20px] max-w-[1280px] m-auto bg-white">
+  <div class="md:mb-[24px] border-r md:border-r-0 sm:bg-white md:bg-white">
     <profile name="<%= @user.name %>"
              display-name="<%= @user.display_name %>"
              phone="<%= @user.phone %>"
@@ -13,12 +13,14 @@
     ></profile>
   </div>
   <div class="grow px-[20px] py-[36px] md:px-0 sm:py-0">
-    <profile-repo-likes-list
+    <collection name="<%= @user.name %>"></collection>
+    <profile-repo-list
       :model-list="<%= JSON.parse(@models).to_json %>"
       :dataset-list="<%= JSON.parse(@datasets).to_json %>"
       :space-list="<%= JSON.parse(@spaces).to_json %>"
       :code-list="<%= JSON.parse(@codes).to_json %>"
       name="<%= @user.name %>"
-    ></profile-repo-likes-list>
+    ></profile-repo-list>
   </div>
 </div>
+
