+++ config/locales/en_js/shared.js
@@ -17,5 +17,5 @@ export const shared = {
   preview: "Preview",
   edit: "Edit",
   likes: "Likes",
-  hasLikes: "Collected"
+  hasLikes: "Liked"
 }
